service update

+++ src/by/epam/filmstore/service/impl/FilmMakerServiceImpl.java
@@ -5,8 +5,10 @@ import by.epam.filmstore.dao.IFilmMakerDAO;
 import by.epam.filmstore.dao.exception.DAOException;
 import by.epam.filmstore.domain.FilmMaker;
 import by.epam.filmstore.domain.Profession;
+import by.epam.filmstore.domain.dto.PagingListDTO;
 import by.epam.filmstore.service.IFilmMakerService;
 import by.epam.filmstore.service.exception.ServiceException;
+import by.epam.filmstore.service.exception.ServiceIncorrectParamLengthException;
 import by.epam.filmstore.service.exception.ServiceValidationException;
 import by.epam.filmstore.service.util.ServiceValidation;
 import by.epam.filmstore.util.DAOHelper;
@@ -17,16 +19,18 @@ import java.util.List;
  * Created by Olga Shahray on 17.07.2016.
  */
 public class FilmMakerServiceImpl implements IFilmMakerService {
+    private final static String ID_MUST_BE_POSITIVE = "Id must be positive number";
+    private final static String MUST_NOT_BE_EMPTY = "Invalid parameters: Fields must not be empty";
 
     @Override
     public void save(String... params) throws ServiceException {
         if (params.length != 2){
-            throw new ServiceValidationException("Invalid arguments: FilmMakerServiceImpl");
+            throw new ServiceIncorrectParamLengthException("Param length must be 2");
         }
         String name = params[0];
         String prof = params[1];
         if(ServiceValidation.isNullOrEmpty(name, prof)){
-            throw new ServiceValidationException("Invalid arguments: FilmMakerServiceImpl. Fields must not be null");
+            throw new ServiceValidationException(MUST_NOT_BE_EMPTY);
         }
         Profession profession = ServiceValidation.getProfession(prof);
 
@@ -45,13 +49,16 @@ public class FilmMakerServiceImpl implements IFilmMakerService {
 
     @Override
     public void update(int id, String... params) throws ServiceException {
-        if (id <= 0 || params.length != 2){
-            throw new ServiceValidationException("Invalid arguments: FilmMakerServiceImpl");
+        if (ServiceValidation.isNotPositive(id)){
+            throw new ServiceValidationException(ID_MUST_BE_POSITIVE);
+        }
+        if(params.length != 2){
+            throw new ServiceIncorrectParamLengthException("Param length must be 2");
         }
         String name = params[0];
         String prof = params[1];
         if(ServiceValidation.isNullOrEmpty(name, prof)){
-            throw new ServiceException("Invalid arguments: FilmMakerServiceImpl. Fields must not be null");
+            throw new ServiceException(MUST_NOT_BE_EMPTY);
         }
         Profession profession = ServiceValidation.getProfession(prof);
 
@@ -69,6 +76,9 @@ public class FilmMakerServiceImpl implements IFilmMakerService {
 
     @Override
     public boolean delete(int id) throws ServiceException {
+        if (ServiceValidation.isNotPositive(id)){
+            throw new ServiceValidationException(ID_MUST_BE_POSITIVE);
+        }
         IFilmMakerDAO dao = DAOFactory.getMySqlDAOFactory().getIFilmMakerDAO();
         try {
             return DAOHelper.execute(() -> dao.delete(id));
@@ -79,6 +89,9 @@ public class FilmMakerServiceImpl implements IFilmMakerService {
 
     @Override
     public FilmMaker get(int id) throws ServiceException {
+        if (ServiceValidation.isNotPositive(id)){
+            throw new ServiceValidationException(ID_MUST_BE_POSITIVE);
+        }
         IFilmMakerDAO dao = DAOFactory.getMySqlDAOFactory().getIFilmMakerDAO();
         try {
             return DAOHelper.execute(() -> dao.get(id));
@@ -88,10 +101,17 @@ public class FilmMakerServiceImpl implements IFilmMakerService {
     }
 
     @Override
-    public List<FilmMaker> getAll(String order, int limit) throws ServiceException {
+    public PagingListDTO<FilmMaker> getAll(int offset, int count) throws ServiceException {
+        if(offset < 0 || count < 0){
+            throw new ServiceValidationException("Offset and count must not be negative");
+        }
         IFilmMakerDAO dao = DAOFactory.getMySqlDAOFactory().getIFilmMakerDAO();
         try {
-            return DAOHelper.execute(() -> dao.getAll(order, limit));
+            return DAOHelper.execute(() -> {
+                List<FilmMaker> list = dao.getAll(offset, count);
+                int countFM = dao.getCountFilmMakers();
+                return new PagingListDTO<FilmMaker>(countFM, list);
+            });
         } catch (DAOException e) {
             throw new ServiceException(e);
         }
@@ -101,7 +121,7 @@ public class FilmMakerServiceImpl implements IFilmMakerService {
     public List<FilmMaker> getAll() throws ServiceException {
         IFilmMakerDAO dao = DAOFactory.getMySqlDAOFactory().getIFilmMakerDAO();
         try {
-            return DAOHelper.execute(() -> dao.getAll());
+            return DAOHelper.execute(dao::getAll);
         } catch (DAOException e) {
             throw new ServiceException(e);
         }
